[sycl-post-link] Remove padding from DeviceGlobalProperty
As it is described in [1], MemorySanitizer emits an unexpected report
about an uninitialized value when we are reading a structure with
padding as a byte array. The patch implements a workaround:
the type of the DeviceImageScope flag has been replaced with uint32_t
to not require a space for padding. The change should have no impact
on performance and ABI because the size is the same.

[1] https://github.com/llvm/llvm-project/issues/54476

+++ llvm/tools/sycl-post-link/DeviceGlobals.h
@@ -35,7 +35,7 @@ struct DeviceGlobalProperty {
 
   // Either 1 (true) or 0 (false), telling whether the device global variable
   // was declared with the device_image_scope property.
-  uint8_t DeviceImageScope;
+  uint32_t DeviceImageScope;
 };
 
 using DeviceGlobalPropertyMapTy =
